refactor useFormFields to use prepareFormFields
